+++ dataflow/src/main/java/org/checkerframework/dataflow/expression/ValueLiteral.java
@@ -110,13 +110,13 @@ public class ValueLiteral extends JavaExpression {
   }
 
   @Override
-  public boolean isUnassignableByOtherCode() {
-    return true;
+  public boolean isAssignableByOtherCode() {
+    return false;
   }
 
   @Override
-  public boolean isUnmodifiableByOtherCode() {
-    return true;
+  public boolean isModifiableByOtherCode() {
+    return false;
   }
 
   @Override